refactor(Masthead): Move `Sheet` to directory
