+++ packages/react-component-library/src/components/TopLevelNavigation/Sheet/Sheet.tsx
@@ -6,7 +6,7 @@ import {
   SHEET_PLACEMENT_ARROW_POSITION_MAP,
 } from './constants'
 
-import { FloatingBox } from '../../primitives/FloatingBox'
+import { FloatingBox } from '../../../primitives/FloatingBox'
 import { SheetButtonProps } from './SheetButton'
 import { useSheet } from './useSheet'
